+++ src/commands/access-profile/ExportAccessProfiles.ts
@@ -167,7 +167,7 @@ class AccessProfileExporter extends BaseCSVExporter<AccessProfile> {
                 const itemDto: AccessProfileDto = {
                     name: item.name,
                     // Escape carriage returns in description.
-                    description: item.description?.replaceAll('\r', "\\r").replaceAll('\n', "\\n"),
+                    description: item.description,
                     enabled: item.enabled,
                     requestable: item.requestable,
                     source: {