+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This changelog is following the recommended format by [keepachangelog](https://k
 
 ### Added
 
+- Add "Role Privileged" column in Campaign Status Report (cf. [#128](https://github.com/yannick-beot-sp/vscode-sailpoint-identitynow/issues/128))
+
 ### Changed
 
 ### Fixed